+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Laravel Query Changelog
 
 ## [Unreleased]
+## [2.0.5]
 ### Updated
 - Complete columns in array keys (technically still values) in methods like `create()`, `update()`, etc. (#40)
 - Make sure it properly resolves table in scenario like so: `$user->customer()->create([])` (#41)